power up on every 200 score

+++ game.js
@@ -9,7 +9,8 @@ let piece = new Piece(ctx);
 let currentPiece = new Piece(ctx);
 currentPiece.makeCurrent();
 
-let powerUp = new PowerUp(ctx);
+let powerUp = null;
+let totalPowerUps = 0;
 
 let gameover = false;
 
@@ -21,7 +22,9 @@ function draw(){
         spawnBoard.draw();
         piece.draw();
         currentPiece.draw();
-        powerUp.draw();
+        if (powerUp != null){
+           powerUp.draw();
+        }
     }
 }
 draw();
@@ -37,11 +40,13 @@ document.onmousedown = function(mouse){
                 currentPiece.picked(mouseX,mouseY);
         }
     }
-    if (mouseX > powerUp.x && 
-        mouseX < powerUp.x + powerUp.width() && 
-        mouseY > powerUp.y && 
-        mouseY < powerUp.y + powerUp.height()){
-            powerUp.picked(mouseX,mouseY);
+    if (powerUp != null){
+        if (mouseX > powerUp.x && 
+            mouseX < powerUp.x + powerUp.width() && 
+            mouseY > powerUp.y && 
+            mouseY < powerUp.y + powerUp.height()){
+                powerUp.picked(mouseX,mouseY);
+        }
     }
     draw();
 }
@@ -72,30 +77,39 @@ document.onmouseup = function (mouse) {
             }
             piece = new Piece(ctx);
         }
-        draw();
     }
-    if (powerUp.isPicked){
-        var mouseX = mouse.clientX - canvas.getBoundingClientRect().left;
-        var mouseY = mouse.clientY - canvas.getBoundingClientRect().top;
-        let destroyedBox = board.destroyBlocks(powerUp.shape,mouseX-powerUp.width()/2+powerUp.boxSize/2,mouseY-powerUp.height()/2+powerUp.boxSize/2);
-        scoreBoard.score += destroyedBox;
-        powerUp.drop();
-        console.log(board.wall);
-        if (board.wall == 0){
-            scoreBoard.score += scoreBoard.level * 5;
-            scoreBoard.level++;
-            if (scoreBoard.level % LEVELGRID.length == 1){
-                scoreBoard.score += 1000;
+    if (powerUp != null){
+        if (powerUp.isPicked){
+            var mouseX = mouse.clientX - canvas.getBoundingClientRect().left;
+            var mouseY = mouse.clientY - canvas.getBoundingClientRect().top;
+            let destroyedBox = board.destroyBlocks(powerUp.shape,mouseX-powerUp.width()/2+powerUp.boxSize/2,mouseY-powerUp.height()/2+powerUp.boxSize/2);
+            scoreBoard.score += destroyedBox;
+            powerUp.drop();
+            console.log(board.wall);
+            if (board.wall == 0){
+                scoreBoard.score += scoreBoard.level * 5;
+                scoreBoard.level++;
+                if (scoreBoard.level % LEVELGRID.length == 1){
+                    scoreBoard.score += 1000;
+                }
+                board.setGrid(LEVELGRID[(scoreBoard.level-1) % LEVELGRID.length]);
             }
-            board.setGrid(LEVELGRID[(scoreBoard.level-1) % LEVELGRID.length]);
+
+            if (board.canPlace){
+                draw();
+                gameover = !board.canPlaceCurrentPiece(currentPiece.shape);
+                if (gameover){
+                    console.log("GAMEOVER");
+                }
+                powerUp = null;
+            }
         }
-        draw();
-        gameover = !board.canPlaceCurrentPiece(currentPiece.shape);
-        if (gameover){
-            console.log("GAMEOVER");
-        }
-        draw();
     }
+    if (Math.floor(scoreBoard.score/200) > totalPowerUps){
+        powerUp = new PowerUp(ctx);
+        totalPowerUps++;
+    }
+    draw();
 }
 
 document.onmousemove = function (mouse) {
@@ -106,11 +120,13 @@ document.onmousemove = function (mouse) {
         draw();
         board.checkPlacebale(currentPiece.shape,mouseX-currentPiece.width()/2+currentPiece.boxSize/2,mouseY-currentPiece.height()/2+currentPiece.boxSize/2);
     }
-    if (powerUp.isPicked){
-        var mouseX = mouse.clientX - canvas.getBoundingClientRect().left;
-        var mouseY = mouse.clientY - canvas.getBoundingClientRect().top;
-        powerUp.picked(mouseX,mouseY);
-        draw();
-        board.checkPlacebale(powerUp.shape,mouseX-powerUp.width()/2+powerUp.boxSize/2,mouseY-powerUp.height()/2+powerUp.boxSize/2);
+    if (powerUp != null){
+        if (powerUp.isPicked){
+            var mouseX = mouse.clientX - canvas.getBoundingClientRect().left;
+            var mouseY = mouse.clientY - canvas.getBoundingClientRect().top;
+            powerUp.picked(mouseX,mouseY);
+            draw();
+            board.checkPlacebale(powerUp.shape,mouseX-powerUp.width()/2+powerUp.boxSize/2,mouseY-powerUp.height()/2+powerUp.boxSize/2);
+        }
     }
 }
